Move base test to utils directory
